разделяет js код на модули
